+++ utils.c
@@ -383,10 +383,10 @@ value_bytes(Var v)
 }
 
 const char *
-raw_bytes_to_moobinary(const char *buffer, int buflen)
+raw_bytes_to_moobinary(const char *buffer, size_t buflen)
 {
     static Stream *s = 0;
-    int i;
+    size_t i;
 
     if (!s)
 	s = new_stream(100);
@@ -404,7 +404,7 @@ raw_bytes_to_moobinary(const char *buffer, int buflen)
 }
 
 const char *
-moobinary_to_raw_bytes(const char *binary, int *buflen)
+moobinary_to_raw_bytes(const char *binary, size_t *buflen)
 {
     static Stream *s = 0;
     const char *ptr = binary;
